clinical_report/backup/requirements.txt
streamlit
google-generativeai
pandas
matplotlib
plotly
python-dotenv
pillow
